feat: theWindow types
